refactored to single config and shorter import paths
