+++ readnext/data/s1_read_d3_documents_chunks.py
@@ -9,7 +9,7 @@ Since the documents dataset is too large to process at once, it is split into ch
 
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 
 
 def write_out_chunked_dataframe(dataframe_chunks: list[pd.DataFrame], file_index: int) -> None:
