+++ readnext/data/s2_preprocess_documents_chunks.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 from readnext.data.preprocessing_utils import add_rank
