+++ readnext/data/s3_merge_labels_chunks.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 
 
 def add_labels(df: pd.DataFrame) -> pd.DataFrame:
