+++ readnext/data/s4_combine_documents_labels_chunks.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 
 
 def main() -> None:
