+++ readnext/data/s5_read_d3_authors.py
@@ -6,7 +6,7 @@ on 2022-11-25.
 
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 
 
 def main() -> None:
