+++ readnext/data/s6_merge_authors.py
@@ -5,7 +5,7 @@ author id.
 
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 from readnext.data.preprocessing_utils import add_rank
