+++ readnext/data/s8_combine_documents_authors_labels_citations_chunks.py
@@ -5,7 +5,7 @@ dataframe.
 
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 
 
 def main() -> None:
