+++ readnext/data/s9_filter_most_cited_documents.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from readnext.data.config import DataPaths
+from readnext.config import DataPaths
 
 
 def main() -> None:
