+++ readnext/modeling/citation_models/__init__.py
@@ -1,0 +1,13 @@
+from readnext.modeling.citation_models.base import (
+    compute_n_most_common,
+    compute_values_df,
+    count_common_values_pairwise,
+    lookup_n_most_common,
+)
+
+__all__ = [
+    "compute_n_most_common",
+    "compute_values_df",
+    "count_common_values_pairwise",
+    "lookup_n_most_common",
+]
