--- readnext/modeling/language_models/modeling/__init__.py
@@ -1,7 +1,0 @@
-from readnext.modeling.language_models.modeling.embedder import (
-    BERTEmbedder,
-    FastTextEmbedder,
-    TFIDFEmbedder,
-)
-
-__all__ = ["BERTEmbedder", "FastTextEmbedder", "TFIDFEmbedder"]
