+++ readnext/modeling/language_models/preprocessing/run.py
@@ -3,9 +3,8 @@ import pickle
 import pandas as pd
 import spacy
 
-from readnext.data.config import DataPaths
-from readnext.modeling.config import ModelVersions, ResultsPaths
-from readnext.modeling.language_models.preprocessing import (
+from readnext.config import DataPaths, ModelVersions, ResultsPaths
+from readnext.modeling.language_models import (
     # BERTTokenizer,
     SpacyTokenizer,
     documents_info_from_df,
