+++ readnext/modeling/language_models/preprocessing/tokenizer.py
@@ -6,7 +6,7 @@ from spacy.language import Language
 from spacy.tokens.doc import Doc
 from transformers import BertTokenizerFast
 
-from readnext.modeling.language_models.preprocessing import DocumentsInfo
+from readnext.modeling.language_models import DocumentsInfo
 from readnext.modeling.utils.rich_progress_bars import setup_progress_bar
 
 DocumentTokens: TypeAlias = list[str]
